Switch to using YCbCr colorspace (replaces HSV)
YCbCr method results in the best results according to the papers below. As an experiment, I compared the results between HSV and YCbCr and the results were MUCH MUCH better. Also, research shows that the YCbCr color space mimics the human visual system.

The biggest issue with YCbCr is picking the correct thresholds for the values based on human skin tone. I plan on experimenting with the color picking tool I have to hopefully make this automated.

http://thescipub.com/PDF/jcssp.2015.1090.1098.pdf

http://www.sersc.org/journals/IJSIP/vol8_no5/11.pdf