Remove remnants of HSV code
Setup color profile to use YCbCr as well and return total min/max of the patches

Change color profile to just indicate the min/max thresholds of each band of the YCbCr colorspace. This can be manually set by the person. As a future improvement to the algorithm, one could have the user pick out pieces of their hand and then determine what color thresholds to use based on the tint of their hand instead of having to do it manually.

Currently, the setup patches UI is still in tact but that will be removed in a later version.

Update the filtering done via binary morphological operations done on the segmented image.