Update algorithm for determining convexity defects.
Calculate angle for convexity defects and throw away any values that are greater than 90 degrees. We know based on human anatomy that the angle of the webbing of our fingers with respect to our fingers should be less than 90 degrees.

Count number of finger webs and return the actual finger count